Cleanup and refactor QGstAppSrc
Clean up the code in QGstAppSrc. Use the new QGst* classes,
and move creation of the GstAppSrc into the class by default.
We still need a way to use an externally defined app src for
playbin support in the audio decoder, so still allow that.

Adjust usages of the API in all places.

Change-Id: Ibcad7e4c836ca9572dfc3d8ed17b7f1980913619
Reviewed-by: André de la Rocha <andre.rocha@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>