Fix bugs in QSoundEffect
Those lead to sound effects not being played back. Also, the
audiodecoder example would not close the wav decoder, leading
to invalid sizes defined inside the wav file.

There's still a bug left that causes the playback to be chopped
into chunks, but that seems to be a separate problem inside the
gstreamer code base.

Change-Id: Id432e7ded128bd2cc0253e12bba06a1595900a3d
Reviewed-by: Doris Verria <doris.verria@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>