Clean up QAudioBuffer API and implementation
And tidy up the docs a bit as well.

QAudioBuffer can really be just a class that contains a
QByteArray for the data, a QAudioFormat and the time stamp.
The QAbstractAudioBuffer design was never used and is overkill
for audio data anyway, as we can afford to copy that once.

Change-Id: I4e1e580db6ecd32b8e92e251f145ab3b851b1bd1
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>