Clean up the videoOutput() API
The current API was rather messy, with multiple overloads.
Change this and make the videoOutput a QObject * based property.
To account for QVideoSink, add a videoSink()/setVideoSink()
overload.

This also helps identifying the correct sink in all cases.

Add some code to protect against deletion of the videoOutput
or videoSink from somewhere else.

Adjust autotests and fix the qquickvideooutput autotests.

With this change we are now passing all existing autotests
again on Linux/gstreamer.

Change-Id: I77b9863e88a7863abf27ab465fadd9c24f486502
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>