Fix a deadlock when changing the video output
Changing the videooutput and deleting the mediaplayer
directly afterwards could lead to a deadlock, as the
videosink was blocking on a pad and then calling
a slot on the main thread using a blocking queued
connection. If the main thread was changing the state of
the gst pipeline at the same time, this could lead to
a deadlock.

Avoid it by pausing the pipeline before changing the
video output.

Change-Id: I0d0cbad0ab97a4eb83dd3c66c664c01de3b0ea90
Reviewed-by: Doris Verria <doris.verria@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>