Fix gstreamer state handling in the capture pipeline
Always keep the capture pipeline running. The old code
that tried to restore things to it's old state was buggy,
as state changes can happen asynchronously, and the old
state could be something that wasn't the state we're
targeting. Because of this, the pipeline could end up
in a paused state even if it was supposed to go to playing.

Fixes the missing preview image in the camera example and
a couple of other bugs.

Change-Id: Ic781817d396058d5265d1ff0ab007808a2227390
Reviewed-by: Doris Verria <doris.verria@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>