Rename QAudioInput to QAudioSource
The renaming is a part of some changes coming from an
API review that tries to solve the problem of not having
API symmetry between QCameraInfo/QCamera and QAudioDeviceInfo
where QAudioDevice is missing (so you can't select a nullptr
as the input/output device to disable it).

One issue is that QAudioInput/QAudioOutput are there for reading
and writing low level PCM data from an audio device. They as such
are not quite suitable as classes to use together with QMediaPlayer
or QMediaCaptureSession, as they allow direct writing of data
bypassing the player or capture pipeline.

To solve this, we renamve those classes to QAudioSource/Sink
(making it a bit more symmetric with QVideoSink). That opens up
our API space for a QAudioDevice class avoiding names that feel
like they are doing the same thing.

The patch also renames things on the backend side to keep
API and implementation consistent.

Change-Id: I9f6b4892d3eee4e252cb3782e681b0c3824f27e2
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>