Use QAudioInput in QMediaCaptureSession
Convert the audioInput property to use a pointer to a QAudioInput
in line with the API changes planned. Not setting up a QAudioInput
implies that you do not get any audio.

Move the volume and muted properties over to QAudioInput, they should
really not be part of the QMediaCaptureSession.

Adjust the platform API accordingly and adjust implementations.

The Android adaptation is incomplete, but that was the case before as
well. This needs some refactoring to properly support setting devices
and adjusting volume/muted.

Dynamically adjusting to changes of the audio input device is not
yet implemented on Windows.

Change-Id: I2fb63b9e57908d6dc02130b8d534ef7ac1a1ac85
Reviewed-by: André de la Rocha <andre.rocha@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>