Make setOutputLocation behavior consistent across supported platforms
outputLocation is always set regardless of whether is it valid or not,
unless platform does not support encoding. If the output location
is not writable an errorOccured signal is emitted.

QNX and Android stored outputLocation in the capture session object.
This has been removed and instead a recording(QUrl location) call
is used to provide location information for the recording session.

Reset m_mediaRecorder after deleting in QAndroidCaptureSession::start

Change-Id: Ib1048410945aaf36532eb8606a4f9eb47460daf7
Reviewed-by: Lars Knoll <lars.knoll@qt.io>