Clean up file format and audio profile handling
Remove the ALAC file format as it is actually always
stored in an mp4 container. While the same is true for
AAC, keep the format, as it's often used with it's own
extension.

Rename ASF to WMV and add WMA as a file format as those
usually have different extensions.

Add WMA and WMV codecs, as the corresponding file formats make
little sense without supporting the codecs that are being used.

Clean up the gstreamer implementation and test that audio only
recording works for all combinations.

Change-Id: I7c76add8f02eba646741c6c2a870e8af9b9fc51b
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>