Move decls in relation to #if'ery to get the code to compile
Fixup for commit 634894a8a83b021386da8c2672628b0b26185ef3 which looks
like it only really meant to suppress the role-related part of the
initialization.  Moved one declaration out of the new #if-ery to
before it; and one from before it to within it, the only use of which
was commented out further down.

Change-Id: I595679eda5ccd848de16a7aa2dc7c39fdc3f329c
Reviewed-by: Doris Verria <doris.verria@qt.io>
Reviewed-by: André de la Rocha <andre.rocha@qt.io>