Provide encoderSettings with the applySettings function
setEncoderSettings is always followed by a call to applySettings, which
is called only once for a sequence of setEncoderSettings calls.
Furthermore, QGstreamerMediaEncoder::setEncoderSettings called
applySettings() which is later called again anyway, leading to multiple
unnecessary applySettings calls. The existence of setEncoderSettings
and applySettings at the same time is confusing and unnecessary

Change-Id: Id24d47c65f54a37a449eeb0a8aeea6a550154c67
Reviewed-by: Lars Knoll <lars.knoll@qt.io>