Resolve media format and provide encoder settings with record()
Media format and other settings is implicitly resolved at the platform
backend and the API user cannot get information about what codecs
bitrates and other settings are selected for recording.

With this patch the resolved media is provided to the backend with
the record() function and the backend updates the remaining encoding
parameters.

Change-Id: I4876397ada2c257e59ebe8fccdd7587b2a628232
Reviewed-by: Doris Verria <doris.verria@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>