Use command like interface to control recorder backend
Recording backends use command like interfaces record/pause/stop. Thus,
reflecting it at the platform level simplifies code.

Additionally, the record function takes the encoder settings, which
helps to further simplify the code. The settings should be apllied when
starting the encoder but not before. The only gain of providing settings
before the record command would be a reduction of the recording startup
time. However, it is doubtful that this speedup is observable.

Change-Id: Iab68594fa0763f67d6e82555627bcdaf08e3d9d3
Reviewed-by: André de la Rocha <andre.rocha@qt.io>