Fix the MapData structure to include per plane size information
So far the data only contained information about the total size in
bytes. That does break down for planar formats where the different
planes do not have to be connected in memory. So rather provide
the data size per plane.

Also clean up the API and remove the methods that access the mapped
data without a plane argument, as their semantic is not well defined.

Change-Id: I3fb2c74172c31720e66d9856472e8786e9004afa
Reviewed-by: André de la Rocha <andre.rocha@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>