Release the lock on the current media file when loading empty URL
At present, it is not possible to delete a file that is open by
the media player even after stopping the playback. A new valid URL
must be provided to the media player in order to release the current
file.

With this patch loading an empty URL or an invalid URL will also
release the lock on the currently loaded media file.

Change-Id: Idc20bad1d6002e2fc0b9f401460ae946bab14099
Reviewed-by: André de la Rocha <andre.rocha@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>