Add support for imx8 video conversions
The iMX8 pipeline is somewhat broken as they send the data over in a
format that can't be handled directly by the sinks (although it's
marked as a supported format). To handle this a special imxvideoconvert_g2d
conversion element is required in the pipeline before the sink.

To make this work, we simply check whether the element is available and in
that case assume that it's required and use it.

Change-Id: I68ee38b56ca21ed7d3d8d8134b4c9b443bf6037d
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>