Enable HW accelerated playback on Android
This requires the new support in qsb to replace a shader with
a custom one so that we can have a shader for the external sampler.

It also breaks abstractions a bit, as we now need to pass the
video frame into the method to get the uniform data.

This is because the external sampler shader requires a per frame
transform on the texture coordinates to be used. The code abuses
the colorTransform uniform that is only used for YUV shaders to
pass that transform to the vertex shader.

Change-Id: Ic934d55507356f310f9beb074d7209bf6fcc8fdd
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>