Use AVCaptureDiscoverySession to get available audio inputs on iOS
Querying the available inputs/outputs through the AVAudioSession will
give us a list of AVAudioSessionPortDescription objects, whose UID
differs from and can not be used to create an AVCaptureDevice for a
session (there seems to be no way of getting the capture device id from
it either). So use AVCaptureDeviceDiscoverySession for now, this gives
us valid capture device ids. A limitation of it is that we can query
only for AVCaptureDeviceTypeBuiltInMicrophone types, which does not
give us Bluetooth or USB devices. These devices were not supported on
5.15 either.
This will need more work in properly configuring the AVAudioSession
when using AVCaptureSession and allowing to set preferred input types.

Change-Id: I7c4dd8de4470a9ef68e21880637be0023691257a
Reviewed-by: Lars Knoll <lars.knoll@qt.io>