Fix the direct texture path
The external texture can now be sampled directly. The external matrix
from the SurfaceTexture needs some adjustments to be correct for us.

The slow path can now be toggled via a build time define (handy for
testing).

There is an additional performance improvement in the uniform data
update for the video node (no need to create a QByteArray - and so
malloc - on every frame, we can rather write into the provided one).

Also a crash fix for when opening a new video in e.g. the mediaplayer
example. (querying the externalMatrix() from the videobuffer should
be skipped when there is no buffer due to switching video sources)

That said, there may be some other issues left because switching videos
in the mediaplayer example does crash sometimes still.

Change-Id: I6542782f4c6e32e0ad8fa0276bc8baac6dda7cad
Reviewed-by: Lars Knoll <lars.knoll@qt.io>