Simplify and fix spectrum build structure
It is not obvious why the example did go to quite a length to
make ffpeg a separate dynamic library. Just build it always
as a static library. The code is LGPLv2.1, but that is compatible
with the BSD-3-Clause license of the example.

Also move the actual application out of the sub-app directory
into the top level directory. This is much more intuitive to use.

Pick-to: 6.2
Change-Id: Ib6173ce21dab78d02190fb0b38c90629b41d26cd
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>