Fix windowed playback on gstreamer for second video
Changing the mediaplayer to a second video using gstreamer would
cause it to come up in a separate top level window, not inside the
QVideoWidget. Fix this by reacting to the GstVideoOverlay's message
requesting a window id instead of setting it in advance (as it
apparently gets reset at some point).

Also avoid pausing the pipeline if the change of sink is a noop.

Change-Id: Ic6e84b76dbbb8253768456fc0865c12634245bba
Reviewed-by: André de la Rocha <andre.rocha@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>