GStreamer: Fix a deadlock when changing audio tracks while paused
When trying to change an audio or subtitle track while the gstreamer
pipeline is paused, one could end up in a deadlock, as the seek
ends up not being able to properly flush the pipeline. But not
seeking will leave us with the wrong audio for a few seconds after
restarting playback.

So do the seek when play() is called again if required. That causes
a pipeline flush and the immediately switch to the new track.

Change-Id: Ie16b2eb7336862623ae31c3bb58b27db06f25ef8
Reviewed-by: André de la Rocha <andre.rocha@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>