Android: create camera session by default
CameraSession will be created by default, but not set to captureSession.
That will allow to set Video Sink any time. CameraSession will be set to
captureSession only when it will be needed (when camera device is set).

When camera is cleared and set once again, the same QVideSink will be
reused.

Change-Id: I72035a0d445dc6ecf43f9b2bc206f4b4ffd8f4bf
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>