Fix error messages when trying to get input device details with OpenSL
When trying to get supported channels and sample rates for input
devices using OpenSL, many error messages are thrown because of
incorrect params provided or because of missing RECORD_AUDIO
permission which is required for that api. Opt out for a list
of default channels and sample rates if the latter permission
is not granted.

Also, get rid of ifdef blocks since this code is used on Android
only.

Pick-to: 6.2
Change-Id: Id4b9922259618b785b3e6cfc3ceae261a789ab98
Reviewed-by: Lars Knoll <lars.knoll@qt.io>