Add some fixes to gstreamer audio output handling
Use the regular 'pause pipeline, do changes, play' mechanism to
manipulate the pipeline instead of doing the change while running.
This mechanism is much more reliable, and we can issue a proper
flush of the pipeline to get immediate audio changes.

Also improve the code that sets the QAudioOutput in the gstreamer
media player so that it can dynamically handle changes there as well.

Change-Id: I38ad302ec8c81417f0d920e26ec2db213093cd9a
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: André de la Rocha <andre.rocha@qt.io>