Improve windowed rendering with gstreamer
Correctly respect the aspectRatioMode when calculating the
render rectangle. This works fine on xvimagesink. The vaapisink
doesn't fully respect it (at least with current AMD drivers), but
that is a bug in the vaapi implementation.

Set the fullscreen flag on the sink when going fullscreen, so that
it can apply any possible optimizations.

Clean up the code dealing with brightness correction etc, as that
is currently not used or supported.

Change-Id: I8e4d005901fa1ee59892d647f9359d8edb96664e
Reviewed-by: André de la Rocha <andre.rocha@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>