Move the GstGlContext initialization into the video sink
This makes the context accessible from more places than just
the sink, something that we need to be able to pass it to the
decoder element before we have actually connected the full
pipeline.

Also fix a bug that triggers a critical warning in latest
gstreamer (don't pass GL_API_ANY to gst_gl_context_new_wrapped).

Change-Id: I9703d9d28766a039b8d1493df181266de41def81
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>