Larger rework of the RGB pixel formats
Until now, pixel formats where not declared in an endian independent way.
Format_ARGB32 meant that we casted the data to an array of integers and
interpreted the highest 8 bits as alpha. This is unfortunate, as this adds
endianness into the picture and complicates things.

Change this towards endian independent ordering of the A, R, G and B
components. New names of those formats are in line with the endian
independent formats in QImage, where RGBA8888 indicates that the first byte
is R and the last byte A.

Add a couple of missing RGB formats, so that we have a symmetrical set and
extend our conversion handling and other places to cover those as well.

Refactor all our code to use the new formats and adjust our built-in
converters to QImage to handle them correctly. Also refactor the
SSE/AVX optimized code paths to handle all RGB based pixel formats.

Adjust our shaders to work correctly with the new definitions.

Finally enable all RGB pixel formats for GL accelerated rendering
on gstreamer, as the byte order issues are now solved.

Change-Id: I8927dc01d569233e9868e54b890f0f7c34705a2b
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>