Enable DMA buffer support for gstreamer
DMA buffer support is only enabled if we have an EGLDisplay, ie on X11
you need to run with QT_XCB_GL_INTEGRATION=xcb_egl. The reason is that
the GL context has to be compatible with EGL, so we can't use DMA with
GLX.

This makes it possible to play back a 4K video inside QML with 5% CPU
load (using a debug build of Qt).

Change-Id: Ie10aab8b1e3f547c54b40c761e06dae6846e1700
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>