Fix critial error related to incorrect gstreamer element state
The audio input and output element state is set to PAUSE regadless of
whether the pipeline is set and if set regardless of the pipeline state.
This leaves the elements in incorrect state and leads to critical error
when destroying them.

Change-Id: Ib4fa7b62267aa1bf0a726c7aef8ff203e5e26229
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit 3b27ea340a50bfc983dc6e68b805b0d249b5145e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>