macOS: Fix image capture
When camera was added after the image capture to the session, the
camera on AVFImageCapture wasn't updated when set later. This was
causing a bug where the camera was reported as not ready and as a
result not being ready to capture. Fix by properly updating the camera
upon the cameraChanged signal.

While we're at it, do some refactoring and cleanup.

Fixes: QTBUG-95641
Change-Id: Ia71e29d5e9fbb410d87c5f3d6399518243010fea
Reviewed-by: André de la Rocha <andre.rocha@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit b219b8b0e8ce5db52aac9c0d0555a297e6c173f7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>