macOS: Fix image capture part 2
Change c2e3377dd8d3c041d2777f2c4a785f6ff8866556 introduced a
regression with image capture where camera times out waiting for a
preview frame. Partly fix this by connecting AVFImageCapture's
slot triggered on a new viewfinder's frame, to the AVFCameraSession::
newVideoFrame signal instead of the AVFCameraRenderer, as the image
capture doesn't know when the video output is added to the session.

In addition, add AVFCameraRenderer to the session when a camera is
first attached and update accordingly when a video sink is set.

Pick-to: 6.2
Change-Id: I32645081e9c1d032691ac23f5ef658e252106c68
Reviewed-by: Lars Knoll <lars.knoll@qt.io>