GStreamer add support for recording pause
Settings the PAUSED state on the encodebin is not sufficient to pause
recording as the encoder still produces data. During the PAUSED state
the video/audio buffers must be dropped. After switch back to recording
the presentation timestamps (PTS) stored in the buffers must be adjusted
for the pause duration.

Pick-to: 6.2
Change-Id: Id2f4706dc8ea421a1a25a1b6e22e9315a3ab3622
Reviewed-by: Lars Knoll <lars.knoll@qt.io>