macOS/iOS: Set pixel format to AVCaptureVideoDataOutput
Normally the capture session's videoDataOutput's pixel
format should be the same as the pixel format set to
the camera through setCameraFormat(QCameraFormat) or its best
supported format if no valid format was set.
However, when creating a texture from the video buffer, there
are problems with pixel formats other than 32ARGB, so for now
enforce this format on the AVCaptureVideoDataOutput.

Pick-to: 6.2
Change-Id: I15f128380bf5e4659d34e09595689deb90a55f28
Reviewed-by: Lars Knoll <lars.knoll@qt.io>