tst_qcamerabackend: Fix test case for setting camera format
Remove framerate comparison between the camera's format and the video
buffer's. The QVideoFrame coming from the newVideoFrame signal, doesn't
have a framerate set.
Also, we are not checking reliably for a format mismatch with QTRY_
VERIFY as it will timeout before even processing a frame (when
formatMismatch is 0 upon initialization). Initialize it to -1 and set
the value of formatMismatch to 1 in case of a mismatch, 0 otherwise.
Blacklist test case for macOS and iOS as viewfinder frame  is always
set to be in Format_BGRA8888 for now.

Change-Id: Ia299900fe8f66b0a5613635cc112d14d5ca10b2e
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit ad1bf0d2e8a5b908301f0ab2ad214a62667e4eea)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>