tst_qcamerabackend: Fix test case for setting camera format
Remove framerate comparison between the camera's format and the video
buffer's. The QVideoFrame coming from the newVideoFrame signal, doesn't
have a framerate set.
Also, we are not checking reliably for a format mismatch with QTRY_
VERIFY as it will timeout before even processing a frame (when
formatMismatch is 0 upon initialization). Initialize it to -1 and set
the value of formatMismatch to 1 in case of a mismatch, 0 otherwise.
Blacklist test case for macOS and iOS as viewfinder frame  is always
set to be in Format_BGRA8888 for now.

Pick-to: 6.2
Change-Id: Ia299900fe8f66b0a5613635cc112d14d5ca10b2e
Reviewed-by: Lars Knoll <lars.knoll@qt.io>