HW accelerated video rendering for QWindow
This implements HW accelerated video rendering to a QWindow using
RHI and the textures we get from the video pipeline.

This change was triggered by difficulties to create a full HW
accelerated path for QVideoWidget on gstreamer. The reason was
that the vaapisink that we've been using so far has quite a few
bugs and been deprecated by the gstreamer maintainers.

This will in a follow-up change allow us to unify the video
pipelines for Qt Quick and Widgets, so that we only need to
support one way of doing HW accelerated rendering on the
backend side.

Done-with: Laszlo Agocs <laszlo.agocs@qt.io>
Change-Id: I143e4416dd8b1cec4e5507e2e5af1eb3d37a496c
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit 1c08cf00aad844512af1aab7b0371541864122c9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>