QAudioDecoder: Re-add API for setting desired audio format
Re-add API for setting/getting desired format for the
decoded audio samples. Readd support on all backends with
the exception of Android.

- Add a warning in the docs for lack of support on Android.

- Make the behavior of audioFormat() consistent on all
platforms. It will return an invalid format if no
format is set. Change Windows implementation to do so.

- Fix the documentation for audioFormat(). The format
returned is the one set to the decoder, and the decoded
buffers may not have this format, in the case when it was
set to an invalid one.

- Some small fixes on darwin and Windows.

Partly reverts af1182b5e29afa0cd69f929d54da224019707214.

Fixes: QTBUG-95766
Change-Id: Id1d6b33a6b9d916e272d3b9d9cee78b6f63609f3
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit 185ff0902181f29990c371d0b19b47915b172143)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>