Clean up the QVideoSink API
The next couple of commits will simplify how we handle HW
accelerated rendering by also going through the RHI code
path for QVideoWidget. This implies that we can remove quite
a few properties from QVideoSink, as those were only required
for the old windowed code path.

This does for the moment disable all HW acceleration on
QVideoWidget. Support for that will be re-added in follow-up
changes.

Pick-to: 6.2
Change-Id: I1474f44ed7468eb263fbf4de75f75f9af16b7773
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Doris Verria <doris.verria@qt.io>