macOS/iOS: Fix adding/removing audio input and output to capture session
In macOS/iOS, we are always initializing a default audio capture device
if no audio input is set to the capture session. This is not correct.
If QAudioInput is set to null, that means the session doesn't get any
audio input. Adjust behavior accordingly.
Also, don't remove the old AVCaptureAudioDataOutput from the capture
session when updating audio input devices. But instead remove the
connection between the old input and output. A connection will be
automatically created between the new input and output.
Update AVFCameraSession accordingly if audio inputs/outputs are set
to null.

Pick-to: 6.2
Change-Id: I79490b1434c8af108597897e961028c1db1931cb
Reviewed-by: Lars Knoll <lars.knoll@qt.io>