GstVideoBuffer: Make sure there is a context current for cleanup
Elsewhere we do not need to worry about an OpenGL context being current
because the various QRhi functions do that for us. However, when
destroying, there are cases (e.g. when closing an application window
with a video widget in it) where there is no context current. Force
a makeCurrent via QRhi then.

Pick-to: 6.2
Change-Id: I1024d53b5229aa35530b0783fb44657c66dc16e9
Reviewed-by: Lars Knoll <lars.knoll@qt.io>