Do graphics pipeline setup after the srb
While probably not an issue for many backends (GL, D3D, maybe Metal
too), due to the fact that the pipeline references the shader
resource bindings object, the srb has to be created beforehand.
Otherwise we will certainly encounter issues with Vulkan, where the srb
is backed by some actual native graphics resource, which is then needed
for the creation of the native pipeline object.

Pick-to: 6.2
Change-Id: I69aff0fcc81ccf3ec50f405c6fe82bb94831f3d2
Reviewed-by: Lars Knoll <lars.knoll@qt.io>