macOS: Clean up dead code related to windowed rendering
We're now always rendering through RHI, so remove the dead code that
was doing the rendering into a layer for QVideoWidget.

This breaks subtitle support on QVideoWidget, but it never worked
correctly for QML. So we'll need a solution that handles both in a
follow-up commit.

Change-Id: Ie0837e3c562edaab861cdb392c8a846f7636b4ed
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit 261849216ce07b4ccfbc2b48f9d73d4b5a75052e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>