macOS/iOS: Fix issues with missed frames and flickering on playback
Playing back high resolution videos did in many cases lead to dropped
frames and flickering. The reason was that we created a new Metal or
OpenGL texture cache for each frame, instead of re-using it.

Fix this by moving the texture cache into the AVFVideoSinkInterface
and passing it over to the AVFVideoBuffer instance. That way we
only ever use one cache object. This seems to fix all playback
issues observed.

Pick-to: 6.2
Change-Id: Id1bd622456bbdeb88fad4cf281a37da032d2c321
Reviewed-by: Doris Verria <doris.verria@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>