macOS/iOS: Implement pause/resume during recording
During pause, the asset writer needs to stop writing the buffers and
modify the timestamps of the buffers after resuming.
With this, we can now pause the writer while changing devices during
the recording as this caused the recorded audio and video to be out of
sync.
Pause and resume the writer accordingly when AVFMediaEncoder::pause()/
resume() is called.

Pick-to: 6.2
Change-Id: I7de96974a1238e544bca8d6edd13fe0c93953d2d
Reviewed-by: Lars Knoll <lars.knoll@qt.io>