macOS/iOS: Allow for other output formats than ARGB32
On Apple platforms, we had so far forced the system to always create
ARGB textures for us. In the longer term that is a bit limiting as it
will for example not allow for HDR rendering.

Fix this and allow a larger variety of video formats now that the
underlying problems with the texture cache are solved.

Change-Id: I7aa2e184a28512b98e3cf1e985539ace7fbaf9ed
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit 6b8089f7a81d32579d42058d392b4754618387e6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>