Remove duplicated media storage location implementations
The implementation of media storage location resolver has been
copied with slight modifications to each platfrom depenedent code.
This patch removes these copies and establishes common naming
scheme for all media captures.

The feature to specify the default directory for a given media type
has been removed as it was only used by Android. However, default
directories for media on Android are properly resolved with
QStandardPaths::writableLocation() Android specific implementation.

Using mutex on the filename generator is unncessary as this function
is not access concurrently.

Holding a counter to generate the name for the next recording does
not help to speed up name generation, since the code checks anyway
with the file system if the generated name file does clash with
the existing files. It is better to scan the directory in one go
to check for the name.

Change-Id: Ibf8d91ea1bacdb3cf347e2097bf065ce715c9157
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit 62aaed9c5937377eb0dee66f9cc6f4d8069d7571)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>