macOS/iOS: Fix audio monitoring for the capture session
The audio output added to the capture session on macOS only monitored
audio if there was a recorder attached to the session. This was because
we only rendered the audio buffers after capturing them in the asset
writer. This is wrong, as we want to monitor audio even without a
recorder set.
To do this, define AVFAudioPreviewDelegate that implements the
AVCaptureAudioDataOutputSampleBufferDelegate protocol to capture the
audio buffers and add it to the session when we set an audio output
through setAudioOutput(QPlatformAudioOutput*). Move all audio preview
related functions in this class.

Pick-to: 6.2
Change-Id: I5fa649682ea146cf5d372752c26c3c2613e8ddef
Reviewed-by: Lars Knoll <lars.knoll@qt.io>