Fix orientation of camera frames on Android
Properly adjust orientation of video frames coming from
the camera. Now we show the video of both cameras the
right way in all orientations.

When in portrait mode, swap the width and height of the
video frames we create, to get the correct aspect
ratio of the camera device.

Change-Id: I7168c672867988c419f160c817a380624392744e
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit 18c1261976d7838814468cb1cc31a3858cf363cd)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>