Remove VideoOutput.autoOrientation property
Late API change. This property got overlooked in earlier reviews.
It doesn't make any sense anymore, as the received video frames
are always oriented the same way.

Adjustments for device/screen orientation on mobile happen inside
the camera backend where we rotate the video frames to are aligned
with the device orientation.

Fixes: QTBUG-94845
Pick-to: 6.2 6.2.0
Change-Id: Iaa85bf127e2a3144f9eba9baaf9560a49ff07abb
Reviewed-by: Samuel Mira <samuel.mira@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>