Fix crash in Android Recorder example on permission request
On the first time the app runs, it asks for permission for recording
video as soon it starts and asks for permission to record audio only
when it is needed.
This permission to record audio can make a crash because requesting
permission makes the app go to background which can make android to
remove the camera surfaceTexture.
Fixed by forcing the app to request permission before creating the
surfaceTexture.

Pick-to: 6.2 6.2.0
Change-Id: I920917a6decf25f7b5f78ee8364d2b9403207aa4
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>