GStreamer: add an option to disable media track with setActiveTrack(-1)
Added an additional disconnect input pad on the input selector for each
media type. This pad is selected as an active-pad when active track is
set to -1.

The decoderPadRemove compared src with inputSelector and terminated
immediately as the src element is the decoder in this case. This is
fixed.

Add a wrapper for inputSelector that handles input track and index
mapping internally.

Change-Id: I2d14795d86f7e1c7e9581cf982008e54a66e498c
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit c205b78f711f32c9c8ac91ebf2c3533355153f9e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>