Disable update orientation when recording on Android
In Android, when the orientation is enabled it will restart the preview
Doing so when the mediaRecorder is recording will crash the app when the
recording stops. So this patch adds a flag in qandroidcamerasession to
prevent updating orientation.

Change-Id: Ibe22035ecb39e95481087e53212fb10c4a6153c5
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit ccb199bba91882981341a2539f2dace2d874c967)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>