AVFVideoRenderer: Properly update video output when changing media
When changing the video source of the AVFMediaPlayer, the video frame
was not updated and displayed the last frame of the previous video.
To fix, before setting a new CALayer to the renderer, make sure to
remove the video output and subtitle output from the previous
AVPlayerItem and add them to the updated one.
Also, clear the last video frame from the video output by setting a
null QVideoFrame() to the sink whenever the mediaplayer stops or
reaches end of stream.

Fixes: QTBUG-96368
Change-Id: I74ed056c1ece0db319e857b89eaf5e5168a972a1
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit 59a32742d66a1fe17320acd133922c105ec211f4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>