Fix Playback issue on Android
The android media player did not change when a new source was
set because in QtAndroidMediaPlayer setDataSource, if the state
was not Idle, would fail silently without changing the source.
Now if the state is not Idle it will reset the mediaplayer.

Fixes: QTBUG-96456
Change-Id: I7b06524ac4a341140e08e44cde2d75211f1c7a9e
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit 96a708759938a2f55dea3b6f02cd394334654818)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>