Fail early on unsupported platform/compiler
Currently, QtMultimedia does not work with the MinGW compiler as the
latter lacks proper support for WMF, which is required by the Windows
backend. When QtMultimedia is built with MinGW, a dummy backend is
built, which returns nullptr for everything. Thus, applications built
with it end up crashing in random locations, which leads developers
to think it's due to some bug, instead of lack of support. It seems
better to fail as soon as the backend is instantiated and also provide
some meaningful message instead.

Task-number: QTBUG-95234
Change-Id: I3c773983c979bbc364810a400bf203fb21400ad6
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: Doris Verria <doris.verria@qt.io>
Reviewed-by: Piotr Srebrny <piotr.srebrny@qt.io>
(cherry picked from commit d519452b438f474e1e24ab568b49efde67c1c65b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>