macOS: Fix video rendering with OpenGL based RHI
Linking against Qt Webengine forces the RHI to be OpenGL based on macOS.
This in turn caused video rendering to fail and only render black or
corrupted rectangles.

The underlying reason is that the OpenGL textures we get from the video
pipeline on macOS are GL_TEXTURE_RECTANGLE based while RHI expects
GL_TEXTURE_2D. As the fix for this requires a bit of plumbing in
QtGui, this change works around the problem by going through a CPU based
texture copy. This works, but leads to higher CPU loads and some dropped
frames at least for 4k videos. Work is ongoing to get this fixed properly
for 6.2.1.

Also submit a two related fixes that were found while debugging the
problem. Ask for OpenGL compatible pixel buffers if we're using OpenGL,
and Flush the texture cache regularly (this was done in Qt 5 as well).

Task-number: QTBUG-96406
Pick-to: 6.2 6.2.0
Change-Id: I0a5f8da1d3d885c40e8b65048da9093d685848aa
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>