Fix android media recorder memory leak
According to the MediaRecorder documentation, namely on the state
diagram, a media recorder can only be released if it is on the initial
state. That was not garanteed before and it could lead to a memory leak.

Change-Id: I926954e04fc0954996ab894ffc38bdbba61d0365
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit 6a1e2758838a29be4626b14da2a93ed04c257fb6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>