Disconnect QAudioOutput from capture session and player in destructor
The expected behavior of elements connected to the capture session
and the player is to disconnect automatically when deleted.
This patch adds this behavior to the QAudioOutput object.
Additionally, this patch enables automatic handover of QAudioOutput
between two capture sessions or a capture session and a player.

Change-Id: I75e5135c54bd11b33876db317cf306d9255f38ce
Reviewed-by: André de la Rocha <andre.rocha@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit b4879defbca4594854e5a8e730e1061e217a7b44)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>