Fix black rectangle on mediaplayer widget example
Multiple widgets are being created, but, in case of mobile ui, not all
of them are added to layout. Those that are created with the player as
the parent are added but not correctly positioned or used.
Reordered calls to better understand what was being created and where.
Prevented creation of widgets not being used on mobile.

Fixes: QTBUG-96747
Pick-to: 6.2
Change-Id: I390ea534d7ea7f58b0fb30faac2bf5491d9723dd
Reviewed-by: Lars Knoll <lars.knoll@qt.io>