GStreamer: remove an element from a pipeline before setState(NULL)
If an element is a part of a pipeline bin, the bin tries to synchronize
its state with its own. Thus, after setting the element state to NULL,
the state can be set back to READY, PAUSE, or PLAYING before it is
removed from the pipeline. This patch first removes an element from
the pipeline and then sets its state to NULL to avoid this race.

Pick-to: 6.2
Change-Id: Ied9cd037fabab19682a53c5ab3d0c7335c58d9c9
Reviewed-by: Lars Knoll <lars.knoll@qt.io>